+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 
 ## `main` branch
 
+**MSRV:** 1.70
+
 ### New Features
 
 * [[PR 171]](https://github.com/rsmpi/rsmpi/pull/171) Refactor attributes to safer approach; add `Communicator::set_attr()`.`
+* [[PR 157]](https://github.com/rsmpi/rsmpi/pull/157) Object safety for `Communicator`
 
 ## 0.7.0 (2023-10-21)